+++ contracts/interfaces/ILendingMarket.sol
@@ -735,6 +735,21 @@ interface ILendingMarketErrors {
      * See details about conditions for the penalty interest rate in the comments for the {Loan} struct.
      */
     error PenaltyInterestRateBelowPrimary();
+
+    /**
+     * @dev Thrown when the penalty interest rate is non-zero before the due date.
+     *
+     * It protects from the situation when the loan duration is changed or can be potentially changed in the future,
+     * because the new duration affects the application of the penalty interest rate of the loan.
+     * See details about the application of the penalty interest rate in the comments for the {Loan} struct.
+     *
+     * If you are sure what you are doing, you can overcome this like:
+     *
+     * - first set the penalty interest rate of the loan to zero,
+     * - then execute the protected operation,
+     * - then set the penalty interest rate back to the original value or a corrected one.
+     */
+    error PenaltyInterestRateNonZeroBeforeDue();
 }
 
 /**